+++ app/src/main/java/com/vkochenkov/filmscatalog/App.kt
@@ -1,0 +1,10 @@
+package com.vkochenkov.filmscatalog
+
+import android.app.Application
+
+class App: Application() {
+
+    override fun onCreate() {
+        super.onCreate()
+    }
+}
