+++ app/src/main/java/com/vkochenkov/filmscatalog/data/Film.kt
@@ -1,4 +1,4 @@
-package com.vkochenkov.filmscatalog.model
+package com.vkochenkov.filmscatalog.data
 
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
